+++ lib/Braintree/TransactionSearch.php
@@ -115,7 +115,9 @@ class Braintree_TransactionSearch
             Braintree_Transaction::SETTLED,
             Braintree_Transaction::SETTLING,
             Braintree_Transaction::SUBMITTED_FOR_SETTLEMENT,
-            Braintree_Transaction::VOIDED
+            Braintree_Transaction::VOIDED,
+            Braintree_Transaction::SETTLEMENT_DECLINED,
+            Braintree_Transaction::SETTLEMENT_PENDING
         ));
     }