+++ tests/integration/ClientTokenTest.php
@@ -29,7 +29,7 @@ class Braintree_ClientTokenTest extends PHPUnit_Framework_TestCase
         $encodedClientToken = Braintree_ClientToken::generate();
         $clientToken = base64_decode($encodedClientToken);
         $version = json_decode($clientToken)->version;
-        $this->assertEquals("2", $version);
+        $this->assertEquals(2, $version);
     }
 
     function test_GatewayRespectsVerifyCard()
@@ -46,7 +46,7 @@ class Braintree_ClientTokenTest extends PHPUnit_Framework_TestCase
         ));
         $authorizationFingerprint = json_decode($clientToken)->authorizationFingerprint;
 
-        $response = Braintree_HttpClientApi::post('/client_api/nonces.json', json_encode(array(
+        $response = Braintree_HttpClientApi::post('/client_api/v1/payment_methods/credit_cards.json', json_encode(array(
             "credit_card" => array(
                 "number" => "4000111111111115",
                 "expirationDate" => "11/2099"
@@ -70,7 +70,7 @@ class Braintree_ClientTokenTest extends PHPUnit_Framework_TestCase
         ));
         $authorizationFingerprint = json_decode($clientToken)->authorizationFingerprint;
 
-        $response = Braintree_HttpClientApi::post('/client_api/nonces.json', json_encode(array(
+        $response = Braintree_HttpClientApi::post('/client_api/v1/payment_methods/credit_cards.json', json_encode(array(
             "credit_card" => array(
                 "number" => "4242424242424242",
                 "expirationDate" => "11/2099"
@@ -89,7 +89,7 @@ class Braintree_ClientTokenTest extends PHPUnit_Framework_TestCase
         ));
         $authorizationFingerprint = json_decode($clientToken)->authorizationFingerprint;
 
-        $response = Braintree_HttpClientApi::post('/client_api/nonces.json', json_encode(array(
+        $response = Braintree_HttpClientApi::post('/client_api/v1/payment_methods/credit_cards.json', json_encode(array(
             "credit_card" => array(
                 "number" => "4242424242424242",
                 "expirationDate" => "11/2099"
@@ -122,7 +122,7 @@ class Braintree_ClientTokenTest extends PHPUnit_Framework_TestCase
         ));
         $authorizationFingerprint = json_decode($clientToken)->authorizationFingerprint;
 
-        $response = Braintree_HttpClientApi::post('/client_api/nonces.json', json_encode(array(
+        $response = Braintree_HttpClientApi::post('/client_api/v1/payment_methods/credit_cards.json', json_encode(array(
             "credit_card" => array(
                 "number" => "4242424242424242",
                 "expirationDate" => "11/2099"