+++ tests/integration/SubscriptionTest.php
@@ -39,6 +39,8 @@ class Braintree_SubscriptionTest extends PHPUnit_Framework_TestCase
         $this->assertInstanceOf('DateTime', $subscription->billingPeriodStartDate);
         $this->assertInstanceOf('DateTime', $subscription->billingPeriodEndDate);
         $this->assertInstanceOf('DateTime', $subscription->paidThroughDate);
+        $this->assertInstanceOf('DateTime', $subscription->updatedAt);
+        $this->assertInstanceOf('DateTime', $subscription->createdAt);
     }
 
     function testCreate_withPaymentMethodNonce()
@@ -590,24 +592,32 @@ class Braintree_SubscriptionTest extends PHPUnit_Framework_TestCase
             'planId' => $plan['id'],
             'descriptor' => array(
                 'name' => '123*123456789012345678',
-                'phone' => '3334445555'
+                'phone' => '3334445555',
+                'url' => 'ebay.com'
             )
         ));
         $this->assertTrue($result->success);
         $subscription = $result->subscription;
         $this->assertEquals('123*123456789012345678', $subscription->descriptor->name);
         $this->assertEquals('3334445555', $subscription->descriptor->phone);
+        $this->assertEquals('ebay.com', $subscription->descriptor->url);
         $transaction = $subscription->transactions[0];
         $this->assertEquals('123*123456789012345678', $transaction->descriptor->name);
         $this->assertEquals('3334445555', $transaction->descriptor->phone);
+        $this->assertEquals('ebay.com', $transaction->descriptor->url);
     }
 
     function testCreate_withDescriptorValidation()
     {
+        $creditCard = Braintree_SubscriptionTestHelper::createCreditCard();
+        $plan = Braintree_SubscriptionTestHelper::addOnDiscountPlan();
         $result = Braintree_Subscription::create(array(
+            'paymentMethodToken' => $creditCard->token,
+            'planId' => $plan['id'],
             'descriptor' => array(
                 'name' => 'xxxxxx',
-                'phone' => 'xxxx'
+                'phone' => 'xxxx',
+                'url' => '12345678901234'
             )
         ));
         $this->assertFalse($result->success);
@@ -618,6 +628,9 @@ class Braintree_SubscriptionTest extends PHPUnit_Framework_TestCase
 
         $errors = $result->errors->forKey('subscription')->forKey('descriptor')->onAttribute('phone');
         $this->assertEquals(Braintree_Error_Codes::DESCRIPTOR_PHONE_FORMAT_IS_INVALID, $errors[0]->code);
+
+        $errors = $result->errors->forKey('subscription')->forKey('descriptor')->onAttribute('url');
+        $this->assertEquals(Braintree_Error_Codes::DESCRIPTOR_URL_FORMAT_IS_INVALID, $errors[0]->code);
     }
 
     function testCreate_fromPayPalACcount()