+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.31.0                                                         
+* Added paypal specific fields to transaction calls               
+* Added SettlementPending, SettlementDeclined transaction statuses
+
+## 2.30.0
+* Add descriptor url support
+
+## 2.29.0
+* Allow credit card verification options to be passed outside of the nonce for PaymentMethod.create
+* Allow billing_address parameters and billing_address_id to be passed outside of the nonce for PaymentMethod.create
+* Add Subscriptions to paypal accounts
+* Add PaymentMethod.update
+* Add fail_on_duplicate_payment_method option to PaymentMethod.create
+
 ## 2.28.0
 * Adds support for v.zero SDKs.